for-loops in mceq barr; PR#677 from PISA master
Tentative merge of PISA master PR https://github.com/icecube/pisa/pull/677: Modified mceq barr stage to use for-loops.
To be discussed:

If I understand correctly, normalizing is not necessary when using the honda flux, as is done in in this fork:
  https://github.com/ts4051/pisa/blob/oscNext2/pisa/stages/flux/mceq_barr.py#L274
This normalization is done in PISA master, independently of whether honda flux is used. This seems unnecessary and should presumably be corrected in PISA master, is that correct?
  https://github.com/icecube/pisa/blob/master/pisa/stages/flux/mceq_barr.py#L278

The differentiation of the nominal_flux_key between neutrino and antineutrino is existing in this fork as well as in the PISA master mceq_barr stage but not in PISA master mceq_barr. Do I understand correctly that this would need to be added to PISA master?
  in this fork: https://github.com/ts4051/pisa/edit/oscNext2/pisa/stages/flux/mceq_barr.py#L411
  in PISA master: https://github.com/icecube/pisa/blob/master/pisa/stages/flux/mceq_barr.py#L438
  in PISA master, mceq_barr_red: https://github.com/icecube/pisa/blob/master/pisa/stages/flux/mceq_barr_red.py#L389